MD5: シンプルなテストの作成
