+++ Makefile
@@ -6,9 +6,11 @@ FILES	:=	\
 			md5.c\
 			md5_block_padding.c\
 			md5_block_rounds.c\
+			md5_derive_digest.c\
 			sha_2.c\
 			sha_256_block_padding.c\
 			sha_256_block_rounds.c\
+			sha_256_derive_digest.c\
 			subbyte_manipulation.c\
 			utils_endian.c\
 
@@ -55,6 +57,17 @@ fclean:			clean
 
 .PHONY:	re
 re:				fclean all
+
+.PHONY:	test_md5
+test_md5: $(NAME)
+	bash test/md5_simple.sh ""
+	bash test/md5_simple.sh "h"
+	bash test/md5_simple.sh "he"
+	bash test/md5_simple.sh "hel"
+	bash test/md5_simple.sh "hell"
+	bash test/md5_simple.sh "hello"
+	bash test/md5_simple.sh "hello!"
+	bash test/md5_simple.sh "974906r9065098t6089569067062078y78yf97t80o7t0o10to8hrr7hgf0o78t872t587f2578fy8gor875boxd9ygn3d098ygbg8on63hcrbuchnyrouifxr7bgoy78b897b8oao78yb789y789bo29n578yo5y89o"
 
 .PHONY:	up
 up:
