+++ srcs/sha_2.c
@@ -11,28 +11,12 @@ static bool	block_operation(t_sha_256_state* state) {
 	// 演算
 	sha_256_block_rounds(state);
 
+	// ブロックバッファをクリア
 	ft_memset(state->schedule.W, 0, sizeof(state->schedule.W));
 	DEBUGWARN("END BLOCK OPERATION for [" U64T ", " U64T ")", state->block_from, state->block_from + SHA_2_WORD_BLOCK_BIT_SIZE);
 	state->block_from += SHA_2_WORD_BLOCK_BIT_SIZE;
 	return state->block_from < state->padded_message_len;
 }
-
-static t_sha_256_digest	derive_digest(const t_sha_256_state* state) {
-	t_sha_256_digest	digest = {};
-	sha_256_word_t		H[8];
-	ft_memcpy(H, state->H, sizeof(H));
-	size_t		digest_index = 0;
-	for (size_t i = 0; i < sizeof(H) / sizeof(sha_256_word_t); ++i) {
-		H[i] = PASS_BIG_END(H[i]);
-		for (size_t j = 0; j < sizeof(sha_256_word_t); ++j) {
-			digest.digest[digest_index] = H[i] % 256;
-			H[i] /= 256;
-			digest_index += 1;
-		}
-	}
-	return digest;
-}
-
 
 t_sha_256_digest	sha_256_hash(const uint8_t* message, uint64_t message_len) {
 	// 状態の初期化
@@ -47,12 +31,11 @@ t_sha_256_digest	sha_256_hash(const uint8_t* message, uint64_t message_len) {
 	// バッチ処理
 	while (block_operation(&state));
 
-	return derive_digest(&state);
+	return sha_256_derive_digest(&state);
 }
 
 void	digest_sha_2(const uint8_t* message, size_t bit_len) {
 	t_sha_256_digest digest = sha_256_hash(message, bit_len);
-	printf("digest:\n");
 	for (size_t i = 0; i < sizeof(digest.digest) / sizeof(uint8_t); i++) {
 		printf("%02x", digest.digest[i]);
 	}
